bench(harbor-adapter): split exit classification into exit_status.py
__init__.py sits at its file-size ceiling; the two exception subclasses,
their exit-code constants, and the verification-failure discriminator are a
coherent seam and move to a sibling module (re-exported, so imports are
unchanged). Unlike exit_cause.py this module requires Harbor by design — it
subclasses the official error.

Refs #1524
